move runtime schema tests
